Add NPM package-lock.json
The Yarn package lock file was removed in https://github.com/krzema12/snakeyaml-engine-kmp/pull/187,
but I didn't add the new NPM `package-lock.json`, but it is required.

Re-add the 'relocate lockfile dir' logic.

Enable reportNewPackageLock so KGP creates an error if the lockfile is missing (otherwise it generates a new one and doesn't complain, which obscures the fact that one is missing!)