Remove irrelevant JavaDoc
Strings are never UTF-8 encoded.
Strings are a sequence of Unicode characters.
If these Unicode characters are encoded into a sequence of bytes
or decoded from a sequence of bytes, then encodings like UTF-8 come into play.

Same for BOM, they are just not relevant when we talk about Strings or Readers,
they are only relevant if starting a stream of bytes, otherwise they are just content.