Temporarily disable Wasm tests (#321)
It's related to
https://github.com/krzema12/snakeyaml-engine-kmp/pull/231.

The test fail on Kotlin 2.1.0 and/or Kotest 6.0, and we don't know why
yet. Given that Wasm is still experimental and unstable, we want to
release the library for other targets. That's why I recommend disabling
the tests, and tracking the issue in
https://github.com/krzema12/snakeyaml-engine-kmp/issues/320. Once the
tests are skipped, we'll be able to unblock merging
https://github.com/krzema12/snakeyaml-engine-kmp/pull/231 (I tested this
change with the mentioned PR and CI is green).

This change is made in a separate PR for visibility.