Publish build scans in case of failures (#325)
Helpful with troubleshooting failed builds, especially if they failed on
a runner that one doesn't have available.

A direct motivation for it are failing tests in
https://github.com/krzema12/snakeyaml-engine-kmp/pull/324. However, the
issue is weird there: if I enable build scans in that PR, the tests
pass. Adding the build scans in general is a good thing, so here's a
separate PR.