performance: avoid boxing when making array's copy (#333)
A small optimization to avoid creating an `Integer` object. Method
`getOrNull` must return an object since it can return a `null` value.
Using `getOrElse` will always return a primitive value. It is not much
but it is something (until I figure out what to do with
`StreamReader.update` method)

Related to #266