Fix reading env vars for JS and WASM JS (only NodeJS) (#335)
While working on
https://github.com/krzema12/snakeyaml-engine-kmp/pull/332, I found out
that the implementations for JS and WASM JS are missing. This PR focuses
on adding them. I don't want to add these in the PR that is related
solely to kotlinizing and commonizing the tests to give this fix a
proper visibility in the release notes.

This change was tested within
https://github.com/krzema12/snakeyaml-engine-kmp/pull/332, and the goal
is to merge that PR soon after this one lands.

Note: env vars work only in a non-browser environment.