Synchronize kotest version to 6.0.0.M1 via version catalog (#398)
I see that Renovate managed to bump two packages, and for other
packages, CI is failing. Not sure if it's because the artifacts are
broken, or the changes need to go together. To solve it, let's make
Renovate bump versions of all kotest-related packages at once, for the
sake of consistency.

This change downgrades some kotest-related libraries to 6.0.0.M1.