Treat Kotlin compilation warnings as errors (#587)
When browsing the build logs (e.g.
[here](https://github.com/krzema12/snakeyaml-engine-kmp/actions/runs/18216126100/job/51865697913)),
warnings clutter the log, especially that in a single log, Kotlin
compiles the code for several targets. It's also visible when e.g.
running the tests in the IDE, and is a source of distraction.

Let's make the logs clean and forbid having unaddressed warnings.