Remove builders for settings classes (#589)
Builders are necessary in Java to fill in for named arguments. They look
fairly Java-ish and don't seem to fit a library that is Kotlin-first.
Thanks to this change, the code will be shorter and more idiomatic
Kotlin.

Two downsides to consider:
* one doesn't have to use the named args, then the readability will be
worse (at least until we have
[KEEP-0249](https://github.com/Kotlin/KEEP/blob/main/proposals/KEEP-0439-named-only-parameters.md)
implemented)
* it's not possible to create a builder and then create various settings
instances off of it. One has to create a complete settings object right
away. This is expected to be 99% the case