Generate copy DSL method using KSP (#593)
It's a follow-up to
https://github.com/krzema12/snakeyaml-engine-kmp/pull/592.

The goal is to generate the boilerplate using a Kotlin Symbol Processing
processor (almost like a compiler plugin). Thanks to this, we can test
the processor's logic, and gain higher confidence that the library's
`copy` DSL is implemented correctly.

The processor isn't aware of the YAML library, so potentially it can be
extracted to a separate library one day.