Remove dependency on unpublished KSP processor module (#601)
Fixes https://github.com/krzema12/snakeyaml-engine-kmp/issues/600.

Fixes an issue where version 4.0.0 of this lib cannot be consumed
because of a missing dependency, a local module that isn't published to
Maven Central. This change just removes the need of depending on the
module. It's achieved by moving the KSP annotation to the library, and
parmeterizing the processor with the annotation's FQN.

Why the kind of dependency isn't changed to `compileOnly`: because for
some targets, Gradle requires the dependency to be added as `api`,
defying the purpose of `compileOnly`. For e. g. the JVM, it would work
as expected.